Feature/extrato endpoint

+++ src/Cepedi.Banco.Conta.Compartilhado/Responses/Transacao/CriarTransacaoResponse.cs
@@ -1,2 +1,2 @@
 namespace Cepedi.Banco.Conta.Compartilhado.Responses;
-public record CriarTransacaoResponse(int idTransacao, string dataTransacao, double valorTransacao);
+public record CriarTransacaoResponse(int idTransacao, string dataTransacao, decimal valorTransacao);
